\name{c-methods}
\alias{c}
\alias{c,McmcdbWide-method}
\alias{c-methods}
\title{Methods for function \code{c}}
\arguments{
  \item{x}{First object}

  \item{...}{Other objects to concatenate to \code{x}.}
}
\value{
  An object of the same class as \code{x}.
}
\description{
  Methods for function \code{c}
}

