\name{coef}
\alias{coef}
\title{Return a coefficient vector for "mcmc" objects.}
\arguments{
  \item{FUN}{Function used to calculate the point
  estimate.}
}
\description{
  This method is meant to be equivalent to the "coef"
  method for "mle" objects, in the sense that it returns a
  vector point estimate for each parameter. By default, the
  point estimate returned is the mean.
}

