\name{create_pararrays}
\alias{create_pararrays}
\title{Create McmcPararrayList from McmcFlatparList}
\usage{
  create_pararrays(x)
}
\arguments{
  \item{x}{\linkS4class{McmcFlatparList}}
}
\value{
  Object of class \linkS4class{McmcPararrayList}
}
\description{
  Create McmcPararrayList from McmcFlatparList
}
\keyword{internal}

