\docType{data}
\name{epil}
\alias{epil}
\title{Epilepsy Data (BUGS Example Vol. 1)}
\description{
  Data from OpenBUGS Examples, Volume 1, "Epilepsy:
  Repeated measures on Poisson Counts"
}
\references{
  \url{http://www.openbugs.info/Examples/Epil.html}

  \href{http://code.google.com/p/stan/source/browse/#git%2Fsrc%2Fmodels%2Fbugs_examples%2Fvol1%2Fepil}{Stan
  source code repsository}
}
\keyword{datasets}

