\docType{methods}
\name{$-methods}
\alias{$}
\alias{$,Mcmcdb-method}
\alias{$-methods}
\title{Methods for function \code{$}}
\arguments{
  \item{x}{An object.}

  \item{name}{\code{character} Paramter array name.}
}
\description{
  Methods for function \code{$}
}
\seealso{
  See \code{\linkS4class{McmcdbWide}} for examples.

  Other McmcdbWide methods: \code{\link{McmcdbWide-class}},
  \code{\link{[}}, \code{\link{[-methods}},
  \code{\link{[[}}, \code{\link{[[-methods}},
  \code{\link{mcmcdb_chains}},
  \code{\link{mcmcdb_samples_flatpars}},
  \code{\link{mcmcdb_samples_flatpars}},
  \code{\link{mcmcdb_samples_iter}},
  \code{\link{mcmcdb_samples_iter}},
  \code{\link{mcmcdb_samples_long}},
  \code{\link{mcmcdb_samples_long}},
  \code{\link{mcmcdb_samples_pararrays}},
  \code{\link{mcmcdb_samples_pararrays}}
}

