\docType{class}
\name{internal-classes}
\alias{McmcChains}
\alias{McmcChains-class}
\alias{McmcFlatparChains}
\alias{McmcFlatparChains-class}
\alias{McmcFlatparChainsOrNull-class}
\alias{McmcIters}
\alias{McmcIters-class}
\alias{McmcSamples}
\alias{McmcSamples-class}
\alias{internal-classes}
\title{Internal classes}
\description{
  Various classes used internally in this package. These
  are all subclasses of \linkS4class{DataFrameConstr}, i.e.
  \code{data.frame} objects with required columns.
}
\examples{
showClass("McmcSamples")
showClass("McmcChains")
showClass("McmcIters")
showClass("McmcFlatparChains")
}
\keyword{classes}

