\docType{data}
\name{line}
\alias{line}
\title{Simple Linear Regression Example}
\source{
  Spiegelhalter, D.J., Thomas, A., Best, N.G. and Gilks,
  W.R. (1995) BUGS: Bayesian inference using Gibbs
  Sampling, Version 0.5, MRC Biostatistics Unit, Cambridge.
}
\description{
  Sample MCMC output from a simple linear regression model
  given in the BUGS manual. These data also appear in the
  \code{coda}, \code{boa} and \code{rjags} packages.
}
\seealso{
  \code{\link[coda]{line}}
}
\keyword{datasets}

