\name{mcmc2}
\alias{mcmc2}
\title{Thi}
\description{
  Thi
}
\section{Slots
\describe{
\item{\code{indices}}{List of matrices which map each flat parameter name to its parameter array.}
\item{\code{template}}{List of parameter arrays in the same form as they were in the estimation.}
}}{

}

\section{Extends
\describe{
\item{\code{mcmc}}{directly}
}}{

}

