\name{mcmcByIteration}
\alias{mcmcByIteration}
\title{Iterate through mcmc iterations}
\arguments{
  \item{object}{\code{McmcList2} object.}

  \item{data}{\code{function} Data to combine with
  parameters on each iteration.}

  \item{FUN}{\code{function} Function to apply to each
  iteration.}

  \item{...}{Pass to \code{FUN}.}
}
\description{
  Iterate through mcmc iterations
}

