\name{mcmcdb_add_parameters}
\alias{mcmcdb_add_parameters}
\alias{mcmcdb_add_parameters,McmcdbWide,matrix-method}
\title{Methods for function \code{mcmcdb_add_parameters}}
\arguments{
  \item{object}{Object to add parameters to.}

  \item{value}{Parameter values to add to \code{object}.}

  \item{parameters}{An object of class
  \code{McmcdbParameters}.}

  \item{flatpar_chains}{An object of class
  \code{McmcdbFlatparChains}, containing information about
  each flat parameter for each chain, such as initial
  values.}
}
\description{
  Methods for function \code{mcmcdb_add_parameters}
}

