\name{mcmcdb_data-methods}
\alias{mcmcdb_data}
\alias{mcmcdb_data,McmcdbMem-method}
\alias{mcmcdb_data-methods}
\title{Get model data from a Mcmcdb object}
\arguments{
  \item{object}{An object for which the method is defined.}
}
\value{
  \code{list} in which each element is some data object.
}
\description{
  Get the data used in model estimation, if it was saved in
  the object.
}

