\docType{methods}
\name{mcmcdb_drop_iters}
\alias{mcmcdb_drop_iters}
\alias{mcmcdb_drop_iters,McmcdbWide,integer-method}
\alias{mcmcdb_drop_iters,McmcdbWide,numeric-method}
\title{Methods for function \code{mcmcdb_drop_iters}}
\arguments{
  \item{object}{An object containing MCMC samples}

  \item{iter}{Iteration identifiers}
}
\value{
  An object of the same class as \code{object}.
}
\description{
  Drop specified iterations from a \code{Mcmcdb} object.
}
\seealso{
  Other McmcdbWide methods: \code{\link{$}},
  \code{\link{$-methods}}, \code{\link{McmcdbWide-class}},
  \code{\link{[}}, \code{\link{[-methods}},
  \code{\link{[[}}, \code{\link{[[-methods}},
  \code{\link{mcmcdb_chains}},
  \code{\link{mcmcdb_drop_chain}},
  \code{\link{mcmcdb_drop_chain}},
  \code{\link{mcmcdb_drop_chain}},
  \code{\link{mcmcdb_drop_pararray}},
  \code{\link{mcmcdb_drop_pararray}},
  \code{\link{mcmcdb_samples_flatpars}},
  \code{\link{mcmcdb_samples_flatpars}},
  \code{\link{mcmcdb_samples_iter}},
  \code{\link{mcmcdb_samples_iter}},
  \code{\link{mcmcdb_samples_long}},
  \code{\link{mcmcdb_samples_long}},
  \code{\link{mcmcdb_samples_pararrays}},
  \code{\link{mcmcdb_samples_pararrays}}
}

