\name{mcmcdb_drop_pararray}
\alias{mcmcdb_drop_pararray}
\alias{mcmcdb_drop_pararray,McmcdbWide,character-method}
\title{Methods for function \code{mcmcdb_drop_pararray}}
\arguments{
  \item{object}{An object containing MCMC samples}

  \item{pararray}{Names of the parameter arrays}
}
\value{
  An object of the same class as \code{object}.
}
\description{
  Drop specified parameter arrays from a \code{Mcmcdb}
  object.
}
\seealso{
  Other McmcdbWide methods: \code{\link{$}},
  \code{\link{$-methods}}, \code{\link{McmcdbWide-class}},
  \code{\link{[}}, \code{\link{[-methods}},
  \code{\link{[[}}, \code{\link{[[-methods}},
  \code{\link{mcmcdb_chains}},
  \code{\link{mcmcdb_samples_flatpars}},
  \code{\link{mcmcdb_samples_flatpars}},
  \code{\link{mcmcdb_samples_iter}},
  \code{\link{mcmcdb_samples_iter}},
  \code{\link{mcmcdb_samples_long}},
  \code{\link{mcmcdb_samples_long}},
  \code{\link{mcmcdb_samples_pararrays}},
  \code{\link{mcmcdb_samples_pararrays}}
}

