\name{mcmcdb_flatpars-methods}
\alias{mcmcdb_flatpars}
\alias{mcmcdb_flatpars,McmcdbMem-method}
\alias{mcmcdb_flatpars,McmcdbParameters-method}
\alias{mcmcdb_flatpars-method}
\alias{mcmcdb_flatpars-methods}
\title{Get flat parameter names}
\arguments{
  \item{object}{Object}
}
\value{
  Named \code{character} vector. Names are the flat
  parameter names; values are the associated parameter
  arrays.
}
\description{
  Get flat parameter names
}
\seealso{
  Other McmcdbMem methods: \code{\link{mcmcdb_iters}},
  \code{\link{mcmcdb_metadata}},
  \code{\link{mcmcdb_parameters}},
  \code{\link{mcmcdb_parameters-methods}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims-methods}}

  Other McmcdbParameters methods:
  \code{\link{McmcdbParameters-class}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims-methods}},
  \code{\link{mcmcdb_paridx}}, \code{\link{mcmcdb_paridx}},
  \code{\link{mcmcdb_paridx-methods}}
}

