\name{mcmcdb_metadata}
\alias{mcmcdb_metadata}
\alias{mcmcdb_metadata,McmcdbWide-method}
\title{Get metadata from an Mcmcdb object}
\arguments{
  \item{object}{object}
}
\value{
  \code{list}
}
\description{
  Get metadata from an Mcmcdb object
}
\seealso{
  \code{\linkS4class{McmcdbWide}}

  Other McmcdbWide methods: \code{\link{McmcdbWide-class}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars-methods}},
  \code{\link{mcmcdb_iters}},
  \code{\link{mcmcdb_par_indices}},
  \code{\link{mcmcdb_par_indices}},
  \code{\link{mcmcdb_par_indices-methods}},
  \code{\link{mcmcdb_parameters}},
  \code{\link{mcmcdb_pararrays}},
  \code{\link{mcmcdb_pararrays}},
  \code{\link{mcmcdb_pararrays-methods}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims-methods}}
}

