\name{mcmcdb_parameters}
\alias{mcmcdb_parameters}
\alias{mcmcdb_parameters,McmcdbWide-method}
\title{Get parameters data from Mcmcdb object}
\arguments{
  \item{object}{object}
}
\value{
  \linkS4class{McmcParameters}
}
\description{
  Get parameters data from Mcmcdb object
}
\seealso{
  \code{\linkS4class{McmcdbWide}}

  Other get.methods: \code{\link{mcmcdb_chains}},
  \code{\link{mcmcdb_iters}},
  \code{\link{mcmcdb_parchains}}
}

