\name{mcmcdb_rename_chain}
\alias{mcmcdb_rename_chain}
\alias{mcmcdb_rename_chain,McmcdbWide-method}
\title{Methods for function \code{mcmcdb_rename_chain}}
\arguments{
  \item{x}{The object to modify.}

  \item{from}{The vector of chain ids to replace}

  \item{to}{The vector of chain ids replacement values}
}
\value{
  An object of the same class as \code{x}.
}
\description{
  Change \code{chain_id} values in a \code{Mcmcdb} object.
}

