\name{mcmcdb_samples_long}
\alias{mcmcdb_samples_long}
\alias{mcmcdb_samples_long,McmcdbWide-method}
\title{Mcmcdb Object samples (Long form)}
\value{
  \code{matrix}. Columns are the long. Rows are the
  iterations (from all chains).
}
\description{
  Return samples from a Mcmcdb object in a matrix
  (iterations x long).
}
\seealso{
  Other McmcdbWide methods: \code{\link{$}},
  \code{\link{McmcdbWide-class}}, \code{\link{[}},
  \code{\link{[[}}, \code{\link{extract-methods}},
  \code{\link{mcmcdb_aply}},
  \code{\link{mcmcdb_aply-methods}},
  \code{\link{mcmcdb_chains}}, \code{\link{mcmcdb_fply}},
  \code{\link{mcmcdb_fply-methods}},
  \code{\link{mcmcdb_samples_flatpars}},
  \code{\link{mcmcdb_samples_flatpars}}
}

