\name{mcmcdb_unflatten-method}
\alias{mcmcdb_unflatten}
\alias{mcmcdb_unflatten,matrix,McmcParameters-method}
\alias{mcmcdb_unflatten,numeric,McmcParameters-method}
\alias{mcmcdb_unflatten-method}
\title{Unflatten MCMC parameters}
\arguments{
  \item{x}{Flattened parameter values}

  \item{parameters}{Object mapping flattened parameters to
  parameter arrays}
}
\value{
  Named \code{list} of parameter arrays
}
\description{
  Unflatten MCMC parameters
}

