\docType{class}
\name{McmcFlatparList-class}
\alias{McmcFlatparList}
\alias{McmcFlatparList-class}
\title{McmcFlatparList class}
\description{
  A list of \linkS4class{McmcFlatpar} objects.
}
\section{Extends}{
  \describe{ \item{\code{namedList}}{directly} }
}
\seealso{
  \linkS4class{McmcFlatpar}
}
\keyword{internal}

