\docType{methods}
\name{mcmcGetIters-method}
\alias{mcmcGetIters-method}
\alias{mcmcGetIters,mcmc.list-method}
\alias{mcmcGetIters,McmcLong-method}
\title{Methods to get MCMC object chain names}
\arguments{
  \item{object}{MCMC object}
}
\description{
  Methods to get MCMC object chain names
}
\keyword{methods}

