\docType{methods}
\name{mcmcGetParname-method}
\alias{mcmcGetParname-method}
\alias{mcmcGetParnames,McmcLong-method}
\alias{mcmcGetParnames,mcmc.list-method}
\title{Methods to get MCMC object parameter names}
\arguments{
  \item{object}{MCMC object}
}
\description{
  Methods to get MCMC object parameter names
}
\keyword{methods}

