\docType{methods}
\name{McmcList2-methods}
\alias{McmcList2,matrix-method}
\alias{McmcList2,mcmc.list-method}
\alias{McmcList2-methods}
\title{Create \code{McmcList2} objects}
\description{
  Create \code{McmcList2} objects
}
\section{Methods}{
  \describe{
  \item{\code{signature(data="McmcList2")}}{Create from
  \code{mcmc.list}.}
  \item{\code{signature(data="McmcList2")}}{Create from
  \code{matrix}.} }
}
\keyword{method}
\keyword{methods}

