\docType{methods}
\name{McmcParameterMeta-methods}
\alias{McmcParameterMeta,character-method}
\alias{McmcParameterMeta,data.frame-method}
\alias{McmcParameterMeta,factor-method}
\alias{McmcParameterMeta,matrix-method}
\alias{McmcParameterMeta-methods}
\title{Create \code{McmcParameterMeta} objects}
\description{
  Create \code{McmcParameterMeta} objects
}
\keyword{methods}

