\name{mcmcUnlist-method}
\alias{mcmcUnlist,array-method}
\alias{mcmcUnlist,list-method}
\alias{mcmcUnlist,numeric-method}
\alias{mcmcUnlist-method}
\alias{mcmcUnlist-methods}
\title{Methods for function \code{mcmcUnlist}}
\arguments{
  \item{x}{Object}

  \item{name}{\code{character}. Parameter name.}

  \item{fun}{Function used to generate unlisted parameter
  names.}
}
\description{
  Methods for function \code{mcmcUnlist}
}

