\docType{methods}
\name{mean-method}
\alias{mean,McmcLong-method}
\alias{mean,mcmc-method}
\alias{mean,mcmc.list-method}
\alias{mean-method}
\title{Methods for MCMC samples}
\description{
  Calculate the mean for each paramter in MCMC samples.
}
\seealso{
  \code{\link[base]{mean}}
}
\keyword{methods}

