\name{median-methods}
\alias{median,mcmc-method}
\alias{median,mcmc.list-method}
\alias{median-methods}
\title{Methods for function \code{methods} for MCMC samples}
\description{
  Calculate the median for each parameter of an MCMC sample
  object.
}
\section{Methods}{
  \describe{ \item{\code{signature(x="mcmc")}}{Method for
  class \code{mcmc}.}
  \item{\code{signature(x="median")}}{Method for class
  \code{mcmc.list}.} }
}
\seealso{
  \code{\link{median}}
}

