\name{melt}
\alias{melt}
\title{melt mcmc objects into a data.frame}
\value{
  \code{data.frame} with columns \describe{
  \item{\code{iteration}}{\code{integer}}
  \item{\code{parameter}}{\code{factor}}
  \item{\code{value}}{\code{numeric}} }

  \code{data.frame} with columns \describe{
  \item{\code{iteration}}{\code{integer}}
  \item{\code{parameter}}{\code{factor}}
  \item{\code{value}}{\code{numeric}} }
}
\description{
  This melts an \code{mcmc} object into a data frame

  This melts an \code{mcmc} object into a data frame
}

