\docType{class}
\name{misc-classes}
\alias{McmcdbChains}
\alias{McmcdbChains-class}
\alias{McmcdbFlatparChains}
\alias{McmcdbFlatparChains-class}
\alias{McmcdbFlatparChainsOrNull-class}
\alias{McmcdbIters}
\alias{McmcdbIters-class}
\alias{McmcdbSamples}
\alias{McmcdbSamples-class}
\alias{misc-classes}
\title{Miscellaneous classes}
\description{
  Various classes used internally in this package. These
  are all subclasses of \linkS4class{DataFrameConstr}, i.e.
  \code{data.frame} objects with required columns.
}
\examples{
showClass("McmcdbSamples")
showClass("McmcdbChains")
showClass("McmcdbIters")
showClass("McmcdbFlatparChains")
}
\keyword{classes}

