\name{SummaryMcmc4}
\alias{SummaryMcmc4}
\title{MCMC summary class}
\description{
  MCMC summary class
}
\section{Slots

\describe{
\item{\code{.Data}}{\code{list} with elements
"statistics", "quantiles", "start", "end", "thin", and "nchain".}
}}{

}

\section{Extends

\describe{
\item{list}{directly}
\item{oldClass}{directly}
}}{

}

