\name{valid_mcmc_parnames}
\alias{valid_mcmc_parnames}
\title{Check MCMC parameter names}
\usage{
  valid_mcmc_parnames(x, style = "bugs")
}
\arguments{
  \item{x}{\code{character} vector of flat parameter names}

  \item{style}{Eiter \code{"bugs"} or \code{"stan"}, to
  check if the parameters are valid BUGS/JAGS or Stan
  names, respectively.}
}
\value{
  \code{logical}.
}
\description{
  Check flat parameter names to see if they are consistent
  with the parameter names generated by BUGS or Stan.
}
\examples{
valid_mcmc_parnames("beta[1]")
valid_mcmc_parnames("beta[1]", "stan")
valid_mcmc_parnames("beta.1", "stan")
}

