\name{create_pararrays}
\alias{create_pararrays}
\title{Create McmcdbPararrayList from McmcdbFlatparList}
\usage{
  create_pararrays(x)
}
\arguments{
  \item{x}{\linkS4class{McmcdbFlatparList}}
}
\value{
  Object of class \linkS4class{McmcdbPararrayList}
}
\description{
  Create McmcdbPararrayList from McmcdbFlatparList
}
\keyword{internal}

