\name{gelman_diag_split}
\alias{gelman_diag_split}
\title{Gelman-Rubin Diagnostic for Split Single Chain}
\arguments{
  \item{x}{\code{mcmc} object. A single chain.

  The single chain is split in half, and then
  \code{gelman_diag_ragged} is run on the two halves.}
}
\description{
  Gelman-Rubin Diagnostic for Split Single Chain
}

