\docType{class}
\name{internal-classes}
\alias{internal-classes}
\alias{McmcChainIters-class}
\alias{McmcChainItersOrNull-class}
\alias{McmcChains-class}
\alias{McmcParChains-class}
\alias{McmcParChainsOrNull-class}
\alias{McmcSamples-class}
\title{Internal classes}
\description{
  Various classes used internally in \pkg{mcmc4}. These are
  all subclasses of \code{DataFramePlus} with specific
  constraints.
}
\keyword{classes}
\keyword{internal}

