\docType{class}
\name{internal-classes}
\alias{McmcChains-class}
\alias{McmcIters-class}
\alias{McmcParChains-class}
\alias{McmcParChainsOrNull-class}
\alias{McmcSamples-class}
\alias{internal-classes}
\title{Internal classes}
\description{
  Various classes used internally in this package. These
  are all subclasses of \linkS4class{DataFrameConstr}, i.e.
  \code{data.frame} objects with required columns.
}
\examples{
showClass("McmcSamples")
showClass("McmcChains")
showClass("McmcIters")
showClass("McmcParChains")
}
\keyword{classes}

