\docType{methods}
\name{mcmc-methods}
\alias{mcmc}
\alias{mcmc,ANY-method}
\alias{mcmc-method}
\alias{mcmc-methods}
\alias{mcmc,numeric-method}
\alias{mcmc,ts-method}
\title{Create \code{mcmc} objects}
\arguments{
  \item{data}{Input object}
}
\description{
  Create an \code{mcmc} object.
}
\section{Methods}{
  \describe{ \item{\code{signature(data = "ANY")}}{}
  \item{\code{signature(data = "numeric")}}{}
  \item{\code{signature(data = "ts")}}{} }
}
\keyword{methods}

