\name{mcmc}
\alias{mcmc}
\alias{mcmc-methods}
\alias{mcmc,ANY-method}
\alias{mcmc,numeric-method}
\alias{mcmc,ts-method}
\title{Create mcmc objects}
\arguments{
  \item{data}{Input object}
}
\description{
  Create an \code{mcmc} object.
}
\details{
  \describe{ \item{\code{signature(data = "ANY")}}{}
  \item{\code{signature(data = "numeric")}}{}
  \item{\code{signature(data = "ts")}}{} }
}

