\name{mcmc}
\alias{mcmc}
\alias{mcmc_numeric}
\alias{mcmc_ts}
\title{Create mcmc}
\usage{
  mcmc_numeric(data, ...)

  mcmc_ts(data, ...)
}
\arguments{
  \item{data}{Input object}

  \item{...}{passed to next generic}
}
\description{
  Generic function to create mcmc objects from various
  types of inputs.
}

