\name{Mcmc2}
\alias{Mcmc2}
\title{Mcmc Samples Object with metadata about parameters}
\description{
  Mcmc Samples Object with metadata about parameters

  Create Mcmc2 objects
}
\section{Slots
\describe{
\item{\code{indices}}{List of matrices which map each flat parameter name to its parameter array.}
\item{\code{template}}{List of parameter arrays in the same form as they were in the estimation.}
}}{

}

\section{Extends
\describe{
\item{\code{mcmc}}{directly}
}}{

}

