\name{mcmcdb_flatpars-methods}
\alias{mcmcdb_flatpars}
\alias{mcmcdb_flatpars,McmcParameters-method}
\alias{mcmcdb_flatpars-method}
\alias{mcmcdb_flatpars-methods}
\title{Get flat parameter names}
\arguments{
  \item{object}{Object}
}
\value{
  Named \code{character} vector. Names are the flat
  parameter names; values are the associated parameter
  arrays.
}
\description{
  Get flat parameter names
}

