\name{mcmcdb_flatpars-methods}
\alias{mcmcdb_flatpars}
\alias{mcmcdb_flatpars,McmcdbParameters-method}
\alias{mcmcdb_flatpars,McmcdbWide-method}
\alias{mcmcdb_flatpars-method}
\alias{mcmcdb_flatpars-methods}
\title{Get flat parameter names}
\arguments{
  \item{object}{Object}
}
\value{
  Named \code{character} vector. Names are the flat
  parameter names; values are the associated parameter
  arrays.
}
\description{
  Get flat parameter names
}
\seealso{
  Other McmcdbParameters methods:
  \code{\link{McmcdbParameters-class}},
  \code{\link{mcmcdb_par_indices}},
  \code{\link{mcmcdb_par_indices}},
  \code{\link{mcmcdb_par_indices-methods}},
  \code{\link{mcmcdb_pararrays}},
  \code{\link{mcmcdb_pararrays}},
  \code{\link{mcmcdb_pararrays-methods}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims-methods}}

  Other McmcdbWide methods: \code{\link{McmcdbWide-class}},
  \code{\link{mcmcdb_iters}},
  \code{\link{mcmcdb_metadata}},
  \code{\link{mcmcdb_par_indices}},
  \code{\link{mcmcdb_par_indices}},
  \code{\link{mcmcdb_par_indices-methods}},
  \code{\link{mcmcdb_parameters}},
  \code{\link{mcmcdb_pararrays}},
  \code{\link{mcmcdb_pararrays}},
  \code{\link{mcmcdb_pararrays-methods}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims-methods}}
}

