\name{mcmcdb_flatpars-methods}
\alias{mcmcdb_flatpars}
\alias{mcmcdb_flatpars,Mcmcdb-method}
\alias{mcmcdb_flatpars,McmcdbParameters-method}
\alias{mcmcdb_flatpars-method}
\alias{mcmcdb_flatpars-methods}
\title{Get flat parameter names}
\arguments{
  \item{object}{Object}
}
\value{
  Named \code{character} vector. Names are the flat
  parameter names; values are the associated parameter
  arrays.
}
\description{
  Get flat parameter names
}
\seealso{
  Other Mcmcdb methods: \code{\link{mcmcdb_mcpar}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims-methods}},
  \code{\link{mcmcdb_paridx}}, \code{\link{mcmcdb_paridx}},
  \code{\link{mcmcdb_paridx-methods}},
  \code{\link{mcmcdb_samples_chain_flatpars}},
  \code{\link{mcmcdb_samples_chain_flatpars}},
  \code{\link{mcmcdb_samples_chain_iter}},
  \code{\link{mcmcdb_samples_chain_iter}},
  \code{\link{mcmcdb_samples_chain_pararrays}},
  \code{\link{mcmcdb_samples_chain_pararrays}},
  \code{\link{mcmcdb_samples_flatpars_chain}},
  \code{\link{mcmcdb_samples_flatpars_chain}},
  \code{\link{mcmcdb_samples_pararrays_chain}},
  \code{\link{mcmcdb_samples_pararrays_chain}},
  \code{\link{mcmdb_mcpar-methods}}

  Other McmcdbParameters methods:
  \code{\link{McmcdbParameters-class}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims-methods}},
  \code{\link{mcmcdb_paridx}}, \code{\link{mcmcdb_paridx}},
  \code{\link{mcmcdb_paridx-methods}}
}

