\name{mcmcdb_metadata}
\alias{mcmcdb_metadata}
\alias{mcmcdb_metadata,McmcdbMem-method}
\title{Get metadata from an Mcmcdb object}
\arguments{
  \item{object}{object}
}
\value{
  \code{list}
}
\description{
  Get metadata from an Mcmcdb object
}
\seealso{
  \code{\linkS4class{McmcdbMem}}

  Other McmcdbMem methods: \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars-methods}},
  \code{\link{mcmcdb_iters}},
  \code{\link{mcmcdb_parameters}},
  \code{\link{mcmcdb_parameters-methods}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims-methods}}
}

