\name{mcmcdb_metadata}
\alias{mcmcdb_metadata}
\alias{mcmcdb_metadata,McmcdbWide-method}
\title{Get metadata from an Mcmcdb object}
\arguments{
  \item{object}{object}
}
\value{
  \code{list}
}
\description{
  Get metadata from an Mcmcdb object
}
\seealso{
  \code{\linkS4class{McmcdbWide}}

  Other McmcdbWide methods: \code{\link{$}},
  \code{\link{McmcdbWide-class}}, \code{\link{[}},
  \code{\link{[[}}, \code{\link{extract-methods}},
  \code{\link{mcmcdb_aply}},
  \code{\link{mcmcdb_aply-methods}},
  \code{\link{mcmcdb_chains}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars-methods}},
  \code{\link{mcmcdb_fply}},
  \code{\link{mcmcdb_fply-methods}},
  \code{\link{mcmcdb_iters}}, \code{\link{mcmcdb_mcpar}},
  \code{\link{mcmcdb_parameters}},
  \code{\link{mcmcdb_parameters-methods}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims-methods}},
  \code{\link{mcmdb_mcpar-methods}}
}

