\name{mcmcdb_parameters-methods}
\alias{mcmcdb_parameters}
\alias{mcmcdb_parameters,McmcdbMem-method}
\alias{mcmcdb_parameters-methods}
\title{Get parameters of Mcmcdb object}
\arguments{
  \item{object}{object}
}
\value{
  An object of class \code{\linkS4class{McmcdbParameters}}.
}
\description{
  Get parameters of Mcmcdb object
}
\seealso{
  Other McmcdbMem methods: \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars-methods}},
  \code{\link{mcmcdb_iters}},
  \code{\link{mcmcdb_metadata}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims-methods}}
}

