\name{mcmcdb_parameters-methods}
\alias{mcmcdb_parameters}
\alias{mcmcdb_parameters,McmcdbWide-method}
\alias{mcmcdb_parameters-methods}
\title{Get paramters of Mcmcdb object}
\arguments{
  \item{object}{object}
}
\value{
  An object of class \code{\linkS4class{McmcdbParameters}}.
}
\description{
  Get paramters of Mcmcdb object
}
\seealso{
  Other McmcdbWide methods: \code{\link{$}},
  \code{\link{McmcdbWide-class}}, \code{\link{[}},
  \code{\link{[[}}, \code{\link{extract-methods}},
  \code{\link{mcmcdb_aply}},
  \code{\link{mcmcdb_aply-methods}},
  \code{\link{mcmcdb_chains}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars-methods}},
  \code{\link{mcmcdb_fply}},
  \code{\link{mcmcdb_fply-methods}},
  \code{\link{mcmcdb_iters}}, \code{\link{mcmcdb_mcpar}},
  \code{\link{mcmcdb_metadata}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims-methods}},
  \code{\link{mcmdb_mcpar-methods}}
}

