\name{mcmcdb_parameters}
\alias{mcmcdb_parameters}
\alias{mcmcdb_parameters,McmcdbWide-method}
\title{Get parameters data from Mcmcdb object}
\arguments{
  \item{object}{object}
}
\value{
  \linkS4class{McmcdbParameters}
}
\description{
  Get parameters data from Mcmcdb object
}
\seealso{
  \code{\linkS4class{McmcdbWide}}

  Other McmcdbWide methods: \code{\link{McmcdbWide-class}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars-methods}},
  \code{\link{mcmcdb_iters}},
  \code{\link{mcmcdb_metadata}},
  \code{\link{mcmcdb_par_indices}},
  \code{\link{mcmcdb_par_indices}},
  \code{\link{mcmcdb_par_indices-methods}},
  \code{\link{mcmcdb_pararrays}},
  \code{\link{mcmcdb_pararrays}},
  \code{\link{mcmcdb_pararrays-methods}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims-methods}}

  Other get-methods: \code{\link{mcmcdb_chains}},
  \code{\link{mcmcdb_flatpar_chains}},
  \code{\link{mcmcdb_iters}}
}

