\name{mcmcdb_pararrays-methods}
\alias{mcmcdb_pararrays}
\alias{mcmcdb_pararrays,McmcdbParameters-method}
\alias{mcmcdb_pararrays,McmcdbWide-method}
\alias{mcmcdb_pararrays-methods}
\title{Get parameter array names}
\arguments{
  \item{object}{Object}
}
\value{
  Named \code{list} of \code{character} vectors. Names are
  the parameter array names; element values are the
  associated flat parameter names.
}
\description{
  Get parameter array names
}
\seealso{
  Other McmcdbParameters methods:
  \code{\link{McmcdbParameters-class}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars-methods}},
  \code{\link{mcmcdb_par_indices}},
  \code{\link{mcmcdb_par_indices}},
  \code{\link{mcmcdb_par_indices-methods}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims-methods}}

  Other McmcdbWide methods: \code{\link{McmcdbWide-class}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars-methods}},
  \code{\link{mcmcdb_iters}},
  \code{\link{mcmcdb_metadata}},
  \code{\link{mcmcdb_par_indices}},
  \code{\link{mcmcdb_par_indices}},
  \code{\link{mcmcdb_par_indices-methods}},
  \code{\link{mcmcdb_parameters}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims}},
  \code{\link{mcmcdb_pardims-methods}}
}

