\name{mcmcdb_pararrays-methods}
\alias{mcmcdb_pararrays}
\alias{mcmcdb_pararrays,McmcParameters-method}
\alias{mcmcdb_pararrays-methods}
\title{Get parameter array names}
\arguments{
  \item{object}{Object}
}
\value{
  Named \code{list} of \code{character} vectors. Names are
  the parameter array names; element values are the
  associated flat parameter names.
}
\description{
  Get parameter array names
}

