\name{mcmcdb_pardims-methods}
\alias{mcmcdb_pardims}
\alias{mcmcdb_pardims,McmcdbParameters-method}
\alias{mcmcdb_pardims,McmcdbWide-method}
\alias{mcmcdb_pardims-methods}
\title{Get parameter array dimensions}
\arguments{
  \item{object}{Object}
}
\value{
  Named \code{list} of \code{integer} vectors. Names are
  the parameter array names; element values are the
  associated dimensions of the arrays.
}
\description{
  Get parameter array dimensions
}
\seealso{
  Other McmcdbParameters methods:
  \code{\link{McmcdbParameters-class}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars-methods}},
  \code{\link{mcmcdb_paridx}}, \code{\link{mcmcdb_paridx}},
  \code{\link{mcmcdb_paridx-methods}}

  Other McmcdbWide methods: \code{\link{McmcdbWide-class}},
  \code{\link{mcmcdb_chains}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars}},
  \code{\link{mcmcdb_flatpars-methods}},
  \code{\link{mcmcdb_iters}},
  \code{\link{mcmcdb_metadata}},
  \code{\link{mcmcdb_parameters}},
  \code{\link{mcmcdb_parameters-methods}}
}

