\name{mcmcdb_pardims-methods}
\alias{mcmcdb_pardims}
\alias{mcmcdb_pardims,McmcParameters-method}
\alias{mcmcdb_pardims-methods}
\title{Get parameter array dimensions}
\arguments{
  \item{object}{Object}
}
\value{
  Named \code{list} of \code{integer} vectors. Names are
  the parameter array names; element values are the
  associated dimensions of the arrays.
}
\description{
  Get parameter array dimensions
}

