\name{McmcdbWide-methods}
\alias{McmcdbWide}
\alias{McmcdbWide,McarrayList-method}
\alias{McmcdbWide,McmcdbWide-method}
\alias{McmcdbWide,StanSamples-method}
\alias{McmcdbWide,data.frame-method}
\alias{McmcdbWide,matrix-method}
\alias{McmcdbWide,mcarray-method}
\alias{McmcdbWide,mcmc-method}
\alias{McmcdbWide,mcmc.list-method}
\alias{McmcdbWide-methods}
\title{Create \code{McmcdbWide} objects}
\arguments{
  \item{object}{An object for which a method is available.}

  \item{model_name}{\code{character} Model name.}

  \item{model_code}{\code{character} Stan model code.}
}
\value{
  An object of class \code{\linkS4class{McmcdbWide}}
  objects.
}
\description{
  Create \code{\linkS4class{McmcdbWide}} objects.
}

