\docType{methods}
\name{McmcParameters-methods}
\alias{McmcParameters,character-method}
\alias{McmcParameters,data.frame-method}
\alias{McmcParameters,factor-method}
\alias{McmcParameters,matrix-method}
\alias{McmcParameters-methods}
\title{Create \code{McmcParameters} objects}
\description{
  Create \code{McmcParameters} objects
}
\keyword{methods}

