\docType{class}
\name{McmcPararrayList-class}
\alias{McmcPararrayList}
\alias{McmcPararrayList-class}
\title{McmcPararray Class}
\usage{
  McmcPararrayList(...)
}
\description{
  A list of \linkS4class{McmcPararray} objects.
}
\section{Extends}{
  \describe{ \item{\code{namedList}}{directly.} }
}
\seealso{
  \code{\link{McmcPararray}}
}
\keyword{internal}

