\name{rbind2-methods}
\alias{rbind2}
\alias{rbind2,McmcdbWide,McmcdbWide-method}
\alias{rbind2,McmcdbWide,missing-method}
\alias{rbind2-methods}
\title{Methods for function \code{rbind2}}
\arguments{
  \item{x}{First object}

  \item{y}{Object to bind to \code{x}.}
}
\value{
  An object of the same class as \code{x}.
}
\description{
  Methods for function \code{rbind2}
}

